Reorganizes types
