+++ src/directives/on.ts
@@ -1,4 +1,4 @@
-import type { Directive } from ".";
+import type { Directive } from "../types";
 
 export const on:Directive = ({ $el, arg, exp, $store, context, get }) => {//, eventName: string, methodOrFunction: string, $store: Record<string, any>, context:ReactiveElement<typeof $store>) => {
     const eventName = arg!.split(/@|:/)[1] as keyof HTMLElementEventMap
